fix _all_ clippy warnings, return value for ref_cnt usize, add some helper for transmutation support, and try to be const-correct

+++ src/skia/color_space.rs
@@ -271,7 +271,7 @@ pub struct XYZD50Hash(pub u32);
 
 #[cfg(test)]
 impl RefCount for ColorSpace {
-    fn ref_cnt(&self) -> i32 {
+    fn ref_cnt(&self) -> usize {
         self.native().ref_cnt()
     }
 }
